unit tests and unsaved behavior fix
This provides a small set of foundational unit tests. Coverage is not
comprehensive byt does attempt to exercise the common workflows.

Note that this also includes a change which aims to fix an issue where
`Set Cookie` would not be sent after the first request regardless of the
value of `save_unchanged`. This occurs because the prior code relied on
the session store to return the cookie value. However in practice this
only happens the first time the cookie is stored.

Here we correct for this by supply the cookie value retrieved from the
request itself (after proper validation). Now when `save_unchanged` is
set, session data has changed, or the cookie value is `None`, we will
set the cookie.