Search every Cookie header for session cookie
HTTP/2 allows a request to have multiple separate Cookie headers with a
single Cookie in each, as opposed to HTTP/1.1 behaviour with all cookies
concatenated into a single Cookie header. Change the method for
searching the session cookie value in a request to account for multiple
Cookie headers, and for multiple cookies in a single header. Add unit
tests for both scenarios.